Working dynamic library with dynamic load

+++ square.c
@@ -1,7 +1,7 @@
 //Функции для integer
 #pragma GCC diagnostic ignored "-Wimplicit-function-declaration"
 #pragma GCC diagnostic ignored "-Wbuiltin-declaration-mismatch"
-int squaring(int num)
+double squaring(int num)
 {
     return (num*num);
 }
